+++ contracts/src/token/erc721/extensions/mod.rs
@@ -7,6 +7,7 @@ pub mod uri_storage;
 pub mod wrapper;
 
 pub use burnable::IErc721Burnable;
+pub use consecutive::Erc721Consecutive;
 pub use enumerable::{Erc721Enumerable, IErc721Enumerable};
 pub use metadata::{Erc721Metadata, IErc721Metadata};
 pub use uri_storage::Erc721UriStorage;